// apps/frontend/src/types/api.ts
export interface components {
    schemas: {
// ...
        UserProfile: components["schemas"]["UserBrief"] & {
            families?: {
                /** Format: uuid */
                id?: string;
                name?: string;
                /** @enum {string} */
                role?: "ORGANIZER" | "MEMBER" | "CHILD";
            }[];
        };
        LoginRequest: {
            /** @description 微信 OAuth code 或 wx.login code */
            code: string;
        };
        LoginResponse: {
            /** @example eyJhbG... */
            accessToken?: string;
            /** @example eyJhbG... */
            refreshToken?: string;
            /** @example 7200 */
            expiresIn?: number;
            user?: components["schemas"]["UserBrief"];
        };
        RefreshRequest: {
            refreshToken: string;
        };
        RefreshResponse: {
            accessToken?: string;
            expiresIn?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}